extract scalar_type traits
